templates: split off template function representation into a separate module
